+++ src/main/java/kitchenpos/menus/tobe/application/MenuEventListener.java
@@ -1,14 +1,15 @@
-package kitchenpos.menus.tobe.infra;
+package kitchenpos.menus.tobe.application;
 
+import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.menus.tobe.domain.menu.Menu;
 import kitchenpos.menus.tobe.domain.menu.MenuRepository;
-import kitchenpos.products.infra.ProductPriceChangesEvent;
+import kitchenpos.products.domain.ProductPriceChangedEvent;
 import org.springframework.context.ApplicationListener;
 import org.springframework.stereotype.Component;
 
 @Component
-public class MenuEventListener implements ApplicationListener<ProductPriceChangesEvent> {
+public class MenuEventListener implements ApplicationListener<ProductPriceChangedEvent> {
   private final MenuRepository menuRepository;
 
   public MenuEventListener(MenuRepository menuRepository) {
@@ -16,11 +17,12 @@ public class MenuEventListener implements ApplicationListener<ProductPriceChange
   }
 
   @Override
-  public void onApplicationEvent(ProductPriceChangesEvent event) {
+  public void onApplicationEvent(ProductPriceChangedEvent event) {
     final List<Menu> menus = menuRepository.findAllByProductId(event.getProductId());
+    final BigDecimal productPrice = event.getProductPrice();
 
     for (final Menu menu : menus) {
-      menu.productPriceChanges();
+      menu.productPriceChanges(productPrice);
     }
   }
 }
